basics of cell rendering, cell types, and core
