+++ Project.toml
@@ -4,6 +4,7 @@ authors = ["emmac <emmettgb@gmail.com>"]
 version = "0.1.0"
 
 [deps]
+IPy = "f2c63c6a-782d-45ad-850e-588fc6f0abf6"
 JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
 Revise = "295af30f-e4ad-537b-8983-00126c2a3abe"
 Toolips = "a47e2ad8-fcf7-417c-9746-32b9bdf220bd"
